refactor(admin): extract SelectionList + shared PermissionsTab (#1104)
Part 2 of the admin dedup. Two honest primitives rather than the issue's
single mega-dialog, which would have needed mode flags to span
client-filter vs server-search and flat vs org-scoped vs two-step.

SelectionList — the checkbox multi-select + client-side search body shared
verbatim by the organizations, roles, and both permissions "Add" dialogs.
The caller owns loading and the selected Set; the primitive owns the >N
search box, the list, and the loading/empty/no-match states. Search state
is internal; callers that must reset it on a context switch (permissions'
resource server, roles' org) give the element a React key so it remounts.

PermissionsTab — one implementation of the role and user permissions tabs,
which were 80% identical (452 + 535 lines). A thin per-resource wrapper
fixes the API resource, the reference target, and — for users — an optional
organization scope. The two tab files are now 7 and 12 lines.

The members "Add" dialog is intentionally left alone: its server-side
search-on-demand is a genuinely different shape, and forcing it in is where
the abstraction would have turned into flags.

Behaviour-preserving; verified by driving the real dialogs in jsdom
(radix Selects and all) with a mock dataProvider — the org-scoped and
role paths each assert their exact create payload — plus SelectionList
interaction tests. These are the first tests over any of this code.
tsc + production build clean; full admin suite 38/38.
