+++ apps/admin/src/resources/users/tabs/roles-tab.tsx
@@ -3,8 +3,7 @@ import { useDataProvider, useNotify } from "ra-core";
 import { useParams } from "react-router-dom";
 import { Plus, Trash2 } from "lucide-react";
 import { Button } from "@/components/ui/button";
-import { Checkbox } from "@/components/ui/checkbox";
-import { Input } from "@/components/ui/input";
+import { SelectionList } from "@/components/SelectionList";
 import {
   Dialog,
   DialogContent,
@@ -142,13 +141,11 @@ function AddRoleDialog({ userId, onAdded }: AddRoleDialogProps) {
   const [available, setAvailable] = useState<RoleRecord[]>([]);
   const [selected, setSelected] = useState<Set<string>>(new Set());
   const [loading, setLoading] = useState(false);
-  const [search, setSearch] = useState("");
 
   const handleOpen = async () => {
     setOpen(true);
     setOrgId(GLOBAL_ID);
     setSelected(new Set());
-    setSearch("");
     try {
       const res = await dataProvider.getList<OrganizationRecord>(
         `users/${userId}/organizations`,
@@ -169,7 +166,6 @@ function AddRoleDialog({ userId, onAdded }: AddRoleDialogProps) {
     setOrgId(GLOBAL_ID);
     setSelected(new Set());
     setAvailable([]);
-    setSearch("");
   };
 
   useEffect(() => {
@@ -279,7 +275,6 @@ function AddRoleDialog({ userId, onAdded }: AddRoleDialogProps) {
               onValueChange={(v) => {
                 setOrgId(v);
                 setSelected(new Set());
-                setSearch("");
               }}
             >
               <SelectTrigger>
@@ -296,67 +291,28 @@ function AddRoleDialog({ userId, onAdded }: AddRoleDialogProps) {
                 ))}
               </SelectContent>
             </Select>
-            {available.length > 5 && (
-              <Input
-                placeholder="Search roles"
-                value={search}
-                onChange={(e) => setSearch(e.target.value)}
-              />
-            )}
-            <div className="max-h-72 overflow-auto border rounded-md">
-              {loading ? (
-                <p className="p-4 text-sm text-muted-foreground">Loading…</p>
-              ) : available.length === 0 ? (
-                <p className="p-4 text-sm text-muted-foreground">
+            <SelectionList
+              key={orgId}
+              items={available.map((r) => ({
+                key: r.id,
+                primary: r.name || r.id,
+                secondary: r.description,
+                searchText: [r.name, r.description, r.id]
+                  .filter(Boolean)
+                  .join(" "),
+              }))}
+              selected={selected}
+              onToggle={toggle}
+              loading={loading}
+              emptyMessage={
+                <>
                   This user already has all available roles
                   {orgId === GLOBAL_ID ? " globally" : " in this organization"}.
-                </p>
-              ) : (
-                (() => {
-                  const q = search.trim().toLowerCase();
-                  const filtered = q
-                    ? available.filter((r) =>
-                        [r.name, r.description, r.id]
-                          .filter((v): v is string => Boolean(v))
-                          .some((v) => v.toLowerCase().includes(q)),
-                      )
-                    : available;
-                  if (filtered.length === 0) {
-                    return (
-                      <p className="p-4 text-sm text-muted-foreground">
-                        No matches
-                      </p>
-                    );
-                  }
-                  return (
-                    <ul className="divide-y">
-                      {filtered.map((r) => (
-                        <li key={r.id} className="flex items-start gap-2 p-2">
-                          <Checkbox
-                            id={`role-${r.id}`}
-                            checked={selected.has(r.id)}
-                            onCheckedChange={() => toggle(r.id)}
-                          />
-                          <label
-                            htmlFor={`role-${r.id}`}
-                            className="flex-1 cursor-pointer"
-                          >
-                            <div className="text-sm font-medium">
-                              {r.name || r.id}
-                            </div>
-                            {r.description && (
-                              <div className="text-xs text-muted-foreground">
-                                {r.description}
-                              </div>
-                            )}
-                          </label>
-                        </li>
-                      ))}
-                    </ul>
-                  );
-                })()
-              )}
-            </div>
+                </>
+              }
+              searchPlaceholder="Search roles"
+              idPrefix="role"
+            />
           </div>
           <DialogFooter>
             <Button variant="outline" onClick={handleClose}>
